Generador de numeros aleatorios de dispositivo cuantico listo para probar
