+++ src/qkd/model/qkd-crypto.cc
@@ -128,83 +128,6 @@ QKDCrypto::~QKDCrypto ()
   //NS_LOG_FUNCTION  (this);  
 }
 
-// IMPRIMIR INFORMACION DE LOS DISPOSITIVOS
-static void _printCardsInfo(QuantisDeviceType deviceType)
-{
-
-  try
-  {
-    // Devices count
-    int devicesCount = idQ::Quantis::Count(deviceType);
-    std::cout << "  Found " << devicesCount << " card(s)" << std::endl;
-
-    // Device details
-    for (int i = 0; i < devicesCount; i++)
-    {
-      // Creates a quantis object
-      idQ::Quantis quantis(deviceType, i);
-
-      // Display device info
-      std::cout << "  - Details for device #" << i << std::endl;
-      int driverVersion = quantis.GetDriverVersion();
-      std::cout << "      driver version: " << static_cast<int>(driverVersion / 10)
-           << "." << driverVersion % 10 << std::endl;
-      std::cout << "      core version: " << std::hex << quantis.GetBoardVersion() << std::endl;
-      std::cout << "      serial number: " << std::hex << quantis.GetSerialNumber() << std::endl;
-      std::cout << "      manufacturer: " << std::hex << quantis.GetManufacturer() << std::endl;
-
-      // Display device's modules info
-      for (int j = 0; j < 4; j++)
-      {
-        std::string strMask = "not found";
-        std::string strStatus = "";
-        if (quantis.GetModulesMask() & (1 << j))
-        {
-          strMask = "found";
-          if (quantis.GetModulesStatus() & (1 << j))
-          {
-            strStatus = "(enabled)";
-          }
-          else
-          {
-            strStatus = "(disabled)";
-          }
-        }
-        std::cout << "      module " << j << ": " << strMask << " " << strStatus << std::endl;
-      }
-    }
-  }
-  catch (std::runtime_error &ex)
-  {
-    std::cerr << "Error while getting cards information: " << ex.what() << std::endl;
-  }
-}
-
-// IMPRIMIR INFORMACION DE LOS DISPOSITIVOS
-static void printCardsInfo()
-{
-  std::cout << "Displaying cards info:" << std::endl;
-
-  std::cout << std::endl
-       << "* Searching for PCI devices..." << std::endl;
-  _printCardsInfo(QUANTIS_DEVICE_PCI);
-
-  std::cout << std::endl
-       << "* Searching for USB devices..." << std::endl;
-  _printCardsInfo(QUANTIS_DEVICE_USB);
-}
-
-void QKDCrypto::initialiceQRNG(){
-    printCardsInfo();
-    std::string params = "u0";
-    idQ::random_device rd(params);
-    for (unsigned int i = 0; i < 100; ++i)
-    {
-        std::cout << "  " << rd() << std::dec << std::endl;
-    }
-
-    
-}
 
 std::vector<uint8_t> 
 QKDCrypto::StringToVector(
